move pre-/post-conditions to BaseTest extend Logger for all classes
